indieauth/confirmation/requirements.txt
jinja2
pyjwt